+++ include/memoria/v1/reactor/msvc/msvc_file.hpp
@@ -108,8 +108,8 @@ public:
 
     virtual void close() = 0;
     
-    virtual uint64_t read(uint8_t* buffer, uint64_t offset, uint64_t size) = 0;
-    virtual uint64_t write(const uint8_t* buffer, uint64_t offset, uint64_t size) = 0;
+    virtual size_t read(uint8_t* buffer, uint64_t offset, size_t size) = 0;
+    virtual size_t write(const uint8_t* buffer, uint64_t offset, size_t size) = 0;
     
     virtual size_t process_batch(IOBatchBase& batch, bool rise_ex_on_error = true) = 0;
